Extract common test methods to TestUtils in test module
