+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/service/NviServiceTest.java
@@ -1,39 +1,34 @@
 package no.sikt.nva.nvi.common.service;
 
+import static no.sikt.nva.nvi.test.TestUtils.extractNviInstitutionIds;
+import static no.sikt.nva.nvi.test.TestUtils.generatePublicationId;
+import static no.sikt.nva.nvi.test.TestUtils.generateS3BucketUri;
+import static no.sikt.nva.nvi.test.TestUtils.mapToVerifiedCreators;
+import static no.sikt.nva.nvi.test.TestUtils.randomPublicationDate;
+import static no.sikt.nva.nvi.test.TestUtils.toPublicationDate;
 import static no.unit.nva.testutils.RandomDataGenerator.randomElement;
-import static no.unit.nva.testutils.RandomDataGenerator.randomLocalDate;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 import static no.unit.nva.testutils.RandomDataGenerator.randomUri;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import java.net.URI;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Stream;
-import no.sikt.nva.nvi.common.model.dao.ApprovalStatus;
 import no.sikt.nva.nvi.common.model.dao.Candidate;
 import no.sikt.nva.nvi.common.model.dao.Level;
-import no.sikt.nva.nvi.common.model.dao.PublicationDate;
-import no.sikt.nva.nvi.common.model.dao.Status;
-import no.sikt.nva.nvi.common.model.dao.VerifiedCreator;
 import no.sikt.nva.nvi.common.model.dto.CandidateDetailsDto;
 import no.sikt.nva.nvi.common.model.dto.EvaluatedCandidateDto;
 import no.sikt.nva.nvi.common.model.dto.PublicationDateDto;
 import no.sikt.nva.nvi.common.model.dto.VerifiedCreatorDto;
-import nva.commons.core.Environment;
-import nva.commons.core.paths.UriWrapper;
+import no.sikt.nva.nvi.test.TestUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class NviServiceTest {
 
-    public static final String BUCKET_HOST = "example.org";
     public static final String CANDIDATE = "Candidate";
-    private static final String PUBLICATION_API_PATH = "publication";
-    private static final Environment ENVIRONMENT = new Environment();
-    private static final String API_HOST = ENVIRONMENT.readEnv("API_HOST");
+
     private NviService nviService;
 
     private FakeNviCandidateRepository fakeNviCandidateRepository;
@@ -63,38 +58,6 @@ public class NviServiceTest {
                    is(equalTo(Optional.of(expectedCandidate))));
     }
 
-    private static ApprovalStatus createPendingApprovalStatus(URI institutionUri) {
-        return new ApprovalStatus.Builder()
-                   .withStatus(Status.PENDING)
-                   .withInstitutionId(institutionUri)
-                   .build();
-    }
-
-    private static PublicationDateDto randomPublicationDate() {
-        var randomDate = randomLocalDate();
-        return new PublicationDateDto(String.valueOf(randomDate.getYear()),
-                                      String.valueOf(randomDate.getMonthValue()),
-                                      String.valueOf(randomDate.getDayOfMonth()));
-    }
-
-    private static Stream<URI> extractNviInstitutionIds(List<VerifiedCreatorDto> creators) {
-        return creators.stream()
-                   .flatMap(creatorDto -> creatorDto.nviInstitutions().stream())
-                   .distinct();
-    }
-
-    private static PublicationDate toPublicationDate(PublicationDateDto publicationDate) {
-        return new PublicationDate(publicationDate.year(),
-                                   publicationDate.month(),
-                                   publicationDate.day());
-    }
-
-    private static List<VerifiedCreator> mapToVerifiedCreators(List<VerifiedCreatorDto> creatorDtos) {
-        return creatorDtos.stream()
-                   .map(creator -> new VerifiedCreator(creator.id(), creator.nviInstitutions()))
-                   .toList();
-    }
-
     private EvaluatedCandidateDto createEvaluatedCandidateDto(UUID identifier,
                                                               List<VerifiedCreatorDto> verifiedCreators,
                                                               String instanceType, Level randomLevel,
@@ -108,10 +71,6 @@ public class NviServiceTest {
                                                                  verifiedCreators));
     }
 
-    private URI generateS3BucketUri(UUID identifier) {
-        return UriWrapper.fromHost(BUCKET_HOST).addChild(identifier.toString()).getUri();
-    }
-
     private Candidate createExpectedCandidate(UUID identifier, List<VerifiedCreatorDto> verifiedCreators,
                                               String instanceType,
                                               Level level, PublicationDateDto publicationDate) {
@@ -123,15 +82,8 @@ public class NviServiceTest {
                    .withIsApplicable(true)
                    .withPublicationDate(toPublicationDate(publicationDate))
                    .withApprovalStatuses(extractNviInstitutionIds(verifiedCreators)
-                                             .map(NviServiceTest::createPendingApprovalStatus)
+                                             .map(TestUtils::createPendingApprovalStatus)
                                              .toList())
                    .build();
-    }
-
-    private URI generatePublicationId(UUID identifier) {
-        return UriWrapper.fromHost(API_HOST)
-                   .addChild(PUBLICATION_API_PATH)
-                   .addChild(identifier.toString())
-                   .getUri();
     }
 }
