+++ nvi-test/build.gradle
@@ -1,0 +1,10 @@
+plugins {
+    id 'nva.nvi.java-conventions'
+}
+dependencies {
+    implementation project(':nvi-commons')
+    implementation libs.nva.commons.apigateway
+    implementation libs.nva.commons.core
+    testImplementation libs.nva.testutils
+    implementation libs.jackson.databind
+}
